Add files via upload
After changing the my_path and file_type variables to the folder path and file type you wish to organize, this program closes any numbering gaps in a folder. For instance, spam003 and spam15 will become spam1 and spam2.

If you have an assortment of prefixes then you should change the first group of the regex to one of the prefixes you wish to organize.