Add files via upload
multiplicationTable.py can be run from command line and takes an integer argument. This program creates an excel sheet, names it multiplicationTable.xlsx, and fills it with a multiple choice table from 1x1 to NxN. Note if you already have an excel file with this name it will be replaced.